feat: navigation-menu cleanup
